Activity에서 패키지명 및 import 문 수정
