+++ app/src/main/java/com/team/runnershi/feature/home/BottomSheetFragment.kt
@@ -1,4 +1,4 @@
-package com.team.runnershi.home
+package com.team.runnershi.feature.home
 
 import android.content.Intent
 import android.os.Bundle
@@ -6,10 +6,9 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import com.google.android.material.bottomsheet.BottomSheetDialogFragment
-import com.team.runnershi.GoalActivity
-import com.team.runnershi.GoalRunMeActivity
+import com.team.runnershi.feature.run.GoalActivity
+import com.team.runnershi.feature.runalone.GoalRunMeActivity
 import com.team.runnershi.R
-import com.team.runnershi.runalone.WaitMeActivity
 import kotlinx.android.synthetic.main.fragment_bottom_sheet.*
