+++ app/src/main/java/com/team/runnershi/feature/home/HomePagerAdapter.kt
@@ -1,12 +1,12 @@
-package com.team.runnershi.adapter
+package com.team.runnershi.feature.home
 
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentManager
 import androidx.fragment.app.FragmentPagerAdapter
-import com.team.runnershi.myprofile.MyProfileFragment
-import com.team.runnershi.record_badge.RecBadgeFragment
-import com.team.runnershi.home.HomeFragment
-import com.team.runnershi.rank.RankFragment
+import com.team.runnershi.feature.myprofile.MyProfileFragment
+import com.team.runnershi.feature.record_badge.RecBadgeFragment
+import com.team.runnershi.feature.home.HomeFragment
+import com.team.runnershi.feature.rank.RankFragment
 
 class HomePagerAdapter(fm : FragmentManager) : FragmentPagerAdapter(fm, BEHAVIOR_RESUME_ONLY_CURRENT_FRAGMENT){
     override fun getItem(position: Int): Fragment {
