+++ app/src/main/java/com/team/runnershi/feature/matchfail/MatchFailActivity.kt
@@ -1,20 +1,19 @@
-package com.team.runnershi.matchfail
+package com.team.runnershi.feature.matchfail
 
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.util.Log
-import com.team.runnershi.rundummy.MatchDummyActivity
+import com.team.runnershi.feature.rundummy.MatchDummyActivity
 import com.team.runnershi.R
 import com.team.runnershi.data.RecordRunWithmeData
 import com.team.runnershi.data.RequestRecordRunWithme
-import com.team.runnershi.data.RequestRecordRunWithmeAlone
 import com.team.runnershi.extension.customEnqueue
 import com.team.runnershi.extension.logDebug
 import com.team.runnershi.extension.newStartActivity
-import com.team.runnershi.home.HomeActivity
+import com.team.runnershi.feature.home.HomeActivity
 import com.team.runnershi.network.RequestToServer
-import com.team.runnershi.runalone.WaitMeActivity
+import com.team.runnershi.feature.runalone.WaitMeActivity
 import com.team.runnershi.util.PrefInit.Companion.prefs
 import kotlinx.android.synthetic.main.activity_match_fail.*
