+++ app/src/main/java/com/team/runnershi/feature/rank/RvRankHonorViewHolder.kt
@@ -1,14 +1,13 @@
-package com.team.runnershi.rank
+package com.team.runnershi.feature.rank
 
 import android.view.View
 import android.widget.ImageView
 import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
 import com.team.runnershi.R
-import com.team.runnershi.data.RunnerData
-import java.math.RoundingMode
+import com.team.runnershi.data.WinnerData
 
-class RvRankMonthViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView){
+class RvRankHonorViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView){
     val rank = itemView.findViewById<TextView>(R.id.tv_rank_rank)
     val profile = itemView.findViewById<ImageView>(R.id.imgv_rank_profile)
     val nickname = itemView.findViewById<TextView>(R.id.tv_rank_nick_name)
@@ -25,12 +24,12 @@ class RvRankMonthViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView){
         R.drawable.icon_redwomen_bunhair
     )
 
-    fun bind(runnerData : RunnerData, ranking: Int){
+    fun bind(winnerData : WinnerData, ranking: Int){
         rank.text =  ranking.toString()
-        profile.setImageResource(imgArray[runnerData.image - 1])
-        nickname.text = runnerData.nickname
+        profile.setImageResource(imgArray[winnerData.image - 1])
+        nickname.text = winnerData.nickname
         //m-> km변환해서 입력
-        dist.text = (runnerData.sum.toDouble()/1000).toBigDecimal().setScale(2, RoundingMode.UP).toDouble().toString()+"km"
+        dist.text = winnerData.win.toString() + "승 " + winnerData.lose.toString() +"패"
     }
 
 }
