+++ app/src/main/java/com/team/runnershi/feature/rank/RvRankLoseViewHolder.kt
@@ -1,18 +1,13 @@
-package com.team.runnershi.rank
+package com.team.runnershi.feature.rank
 
-import android.graphics.Color
-import android.util.DisplayMetrics
 import android.view.View
-import android.view.ViewGroup
 import android.widget.ImageView
 import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
 import com.team.runnershi.R
-import com.team.runnershi.data.RunnerData
-import com.team.runnershi.data.WinnerData
-import org.w3c.dom.Text
+import com.team.runnershi.data.LoserData
 
-class RvRankHonorViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView){
+class RvRankLoseViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView){
     val rank = itemView.findViewById<TextView>(R.id.tv_rank_rank)
     val profile = itemView.findViewById<ImageView>(R.id.imgv_rank_profile)
     val nickname = itemView.findViewById<TextView>(R.id.tv_rank_nick_name)
@@ -29,12 +24,12 @@ class RvRankHonorViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView){
         R.drawable.icon_redwomen_bunhair
     )
 
-    fun bind(winnerData : WinnerData, ranking: Int){
+    fun bind(loserData : LoserData, ranking: Int){
         rank.text =  ranking.toString()
-        profile.setImageResource(imgArray[winnerData.image - 1])
-        nickname.text = winnerData.nickname
+        profile.setImageResource(imgArray[loserData.image - 1])
+        nickname.text = loserData.nickname
         //m-> km변환해서 입력
-        dist.text = winnerData.win.toString() + "승 " + winnerData.lose.toString() +"패"
+        dist.text = loserData.win.toString() + "승 " + loserData.lose.toString() +"패"
     }
 
 }
