+++ app/src/main/java/com/team/runnershi/feature/rank/RvRankMonthViewHolder.kt
@@ -1,19 +1,14 @@
-package com.team.runnershi.rank
+package com.team.runnershi.feature.rank
 
-import android.graphics.Color
-import android.util.DisplayMetrics
 import android.view.View
-import android.view.ViewGroup
 import android.widget.ImageView
 import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
 import com.team.runnershi.R
-import com.team.runnershi.data.LoserData
 import com.team.runnershi.data.RunnerData
-import com.team.runnershi.data.WinnerData
-import org.w3c.dom.Text
+import java.math.RoundingMode
 
-class RvRankLoseViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView){
+class RvRankMonthViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView){
     val rank = itemView.findViewById<TextView>(R.id.tv_rank_rank)
     val profile = itemView.findViewById<ImageView>(R.id.imgv_rank_profile)
     val nickname = itemView.findViewById<TextView>(R.id.tv_rank_nick_name)
@@ -30,12 +25,12 @@ class RvRankLoseViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView){
         R.drawable.icon_redwomen_bunhair
     )
 
-    fun bind(loserData : LoserData, ranking: Int){
+    fun bind(runnerData : RunnerData, ranking: Int){
         rank.text =  ranking.toString()
-        profile.setImageResource(imgArray[loserData.image - 1])
-        nickname.text = loserData.nickname
+        profile.setImageResource(imgArray[runnerData.image - 1])
+        nickname.text = runnerData.nickname
         //m-> km변환해서 입력
-        dist.text = loserData.win.toString() + "승 " + loserData.lose.toString() +"패"
+        dist.text = (runnerData.sum.toDouble()/1000).toBigDecimal().setScale(2, RoundingMode.UP).toDouble().toString()+"km"
     }
 
 }
